@@ -4,45 +4,47 @@
   "description": "",
   "main": "app.js",
   "scripts": {
-    "start": "nodemon",
-    "build": "tsc -b"
+    "dev": "nodemon",
+    "build": "tsc -b",
+    "db:push": "prisma db push",
+    "db:studio": "prisma studio"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@prisma/client": "^5.9.1",
-    "@types/jsonwebtoken": "^9.0.5",
-    "@types/nodemailer": "^6.4.14",
-    "@upstash/redis": "^1.28.3",
+    "@prisma/client": "^5.14.0",
+    "@types/jsonwebtoken": "^9.0.6",
+    "@types/nodemailer": "^6.4.15",
+    "@upstash/redis": "^1.31.1",
     "bcrypt": "^5.1.1",
     "body-parser": "^1.20.2",
     "compression": "^1.7.4",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
-    "dotenv": "^16.4.1",
-    "express": "^4.18.2",
+    "dotenv": "^16.4.5",
+    "express": "^4.19.2",
     "jsonwebtoken": "^9.0.2",
-    "nodemailer": "^6.9.9",
-    "nodemon": "^3.0.3",
-    "pg": "^8.11.3",
-    "uploadthing": "^6.3.3"
+    "nodemailer": "^6.9.13",
+    "nodemon": "^3.1.0",
+    "pg": "^8.11.5",
+    "uploadthing": "^6.10.4"
   },
   "devDependencies": {
     "@types/bcrypt": "^5.0.2",
     "@types/body-parser": "^1.19.5",
     "@types/compression": "^1.7.5",
-    "@types/cookie-parser": "^1.4.6",
+    "@types/cookie-parser": "^1.4.7",
     "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
     "@types/jsonwebtoken": "^9.0.5",
-    "@types/node": "^20.11.16",
+    "@types/node": "^20.12.12",
     "@types/nodemailer": "^6.4.10",
-    "@types/pg": "^8.11.0",
+    "@types/pg": "^8.11.6",
     "nodemon": "^3.0.1",
-    "prisma": "^5.9.1",
+    "prisma": "^5.14.0",
     "ts-node": "^10.9.2",
     "tsconfig-paths": "^4.2.0",
-    "typescript": "^5.3.3"
+    "typescript": "^5.4.5"
   }
 }
