@@ -5,7 +5,7 @@ export const verifyUpdation = async (req: express.Request, res: express.Response
    const {email, oneTimePass} = req.body;
    const otp = Number(oneTimePass)
    const user = await prisma.user.findUnique({ where: { email: email }});
-   if (user.otp === otp) {
+   if (user?.otp === otp) {
      await prisma.user.update({where: {email: email}, data: {otp: null, is_verified: true}});
      res.json({ message: "User Verified", verified: true })
    } else {
