@@ -14,9 +14,9 @@ export const featuredCars = async (req: express.Request, res: express.Response) 
       lastShuffleTime = currentTime;
     }
     const selectedModels = shuffledModels.slice(0, 7);
-    res.json({ total: selectedModels.length, models: selectedModels });
+    res.json({ success: true, total: selectedModels.length, models: selectedModels });
   } catch (error) {
-    res.json({ total: null, models: null });
+    res.json({ success: false, total: null, models: null });
   }
 };
 
