@@ -1,21 +1,12 @@
 import express from "express";
 import prisma from "../../../db/db.config";
+import { UserRequest } from "../../../types/UserRequest";
 
-interface User {
-   email: string;
-   name: string;
-   password: string;
-   is_Verified: boolean;
- }
- interface RequestWithUser extends express.Request {
-   user: User;
- }
-
-export const wishlistCar = async (req: RequestWithUser, res: express.Response) => {
+export const wishlistCar = async (req: UserRequest, res: express.Response) => {
    const id = parseInt(req.params.id);
    const car = await prisma.model.findUnique({ where: { id: id } });
    if (car) {
-     const user = await prisma.user.findUnique({where: {email: req.user.email}});
+     const user = await prisma.user.findUnique({where: {email: req.user?.email}});
      if (user) {
        const userID = user.id;
        const carID = car.id;
