@@ -1,22 +1,13 @@
 import express from "express";
 import prisma from "../../../db/db.config";
+import { UserRequest } from "../../../types/UserRequest";
 
-interface User {
-   email: string;
-   name: string;
-   password: string;
-   is_Verified: boolean;
- }
- interface RequestWithUser extends express.Request {
-   user: User;
- }
-
-export const deleteWishlistedCar = async (req: RequestWithUser, res: express.Response) => {
+export const deleteWishlistedCar = async (req: UserRequest, res: express.Response) => {
    const wishCar = await prisma.wishlistedCar.findUnique({
      where: { id: parseInt(req.params.id) },
    });
    if (wishCar) {
-     const user = await prisma.user.findUnique({where: { email: req.user.email }, include: {onWishlist: true}});
+     const user = await prisma.user.findUnique({where: { email: req.user?.email }, include: {onWishlist: true}});
      if (user) {
        const wishCarId = wishCar.id;
        await prisma.wishlistedCar.delete({where: {id: wishCarId}});
