@@ -5,7 +5,7 @@ import { verifyRegistration } from "../controllers/user/auth/verify-registration
 import { forgotPassword } from "../controllers/user/auth/forgot-password";
 import { verifyUpdation } from "../controllers/user/auth/verify-updation";
 import { resetPassword } from "../controllers/user/auth/reset-password";
-import { inventory } from "../controllers/user/collections/inventory";
+import { inventory } from "../controllers/user/collections/all-cars";
 import { singleCar } from "../controllers/user/collections/single-car";
 import { rentCar } from "../controllers/user/rent/rent-car";
 import { rentStatus } from "../controllers/user/rent/rent-status";
@@ -13,9 +13,9 @@ import { rentedCars } from "../controllers/user/rent/rented-cars";
 import { wishlistCar } from "../controllers/user/wishlist/wishlist-car";
 import { wishlistedCars } from "../controllers/user/wishlist/wishlisted-cars";
 import { deleteWishlistedCar } from "../controllers/user/wishlist/wishlist-delete";
-import { authentication } from "../middleware/userAuthenticator";
 import { totalCars } from "../controllers/user/collections/total-cars";
 import { featuredCars } from "../controllers/user/collections/featured-cars";
+import { authentication } from "../middleware/userAuthenticator";
 
 const router = express.Router();
 
