@@ -1,12 +1,14 @@
-import React from 'react'
-import AppRoutes from '@/routes/AppRoutes'
+import React from 'react';
+import AppRoutes from '@/routes/AppRoutes';
+import { Toaster } from 'react-hot-toast';
 
 const App: React.FC = () => {
-  return (
-    <>
-      <AppRoutes />
-    </>
-  )
-}
+   return (
+      <>
+         <Toaster position="top-right" reverseOrder={false} />
+         <AppRoutes />
+      </>
+   );
+};
 
-export default App
+export default App;
