@@ -37,7 +37,7 @@ const DrawerContent = React.forwardRef<
   React.ComponentPropsWithoutRef<typeof DrawerPrimitive.Content>
 >(({ className, children, ...props }, ref) => (
   <DrawerPortal>
-    <DrawerOverlay />
+    <DrawerOverlay className="DrawerOverlay"/>
     <DrawerPrimitive.Content
       ref={ref}
       className={cn(
