@@ -1,17 +1,13 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import App from './App.tsx'
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import App from './App.tsx';
 import { BrowserRouter } from 'react-router-dom';
-import './index.css'
-import { Provider } from 'react-redux';
-import { store } from "./state/store.ts"
+import './index.css';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
-  <BrowserRouter>
-    <React.StrictMode>
-      <Provider store={store}>
-        <App />
-      </Provider>
-    </React.StrictMode>
-  </BrowserRouter>
-)
+   <BrowserRouter>
+      <React.StrictMode>
+         <App />
+      </React.StrictMode>
+   </BrowserRouter>,
+);
