@@ -1,22 +1,25 @@
-import About from '@/app/contacts/AboutUs'
-import ContactUs from '@/app/contacts/ContactUs'
+import About from '@/app/extra/AboutUs'
+import ContactUs from '@/app/extra/ContactUs'
 import Home from '@/app/home/Home'
-import TheCollection from '@/app/user/TheCollection'
-import Login from '@/app/user/auth/Login'
-import Register from '@/app/user/auth/Register'
-import VerifyRegistation from '@/app/user/auth/VerifyRegistration'
-import AllCars from '@/app/user/cars/inventory/AllCars'
-import RentACar from '@/app/user/cars/inventory/RentACar'
-import SingleCarCard from '@/app/user/cars/inventory/SingleCarCard'
-import Success from '@/app/user/checkout/Success'
-import Wishlist from '@/app/user/wishlist/Wishlist'
+import TheCollection from '@/app/showcase/TheCollection'
+import Login from '@/app/auth/Login'
+import Register from '@/app/auth/Register'
+import VerifyRegistation from '@/app/auth/VerifyRegistration'
+import AllCars from '@/app/showcase/cars/inventory/AllCars'
+import RentACar from '@/app/showcase/cars/inventory/RentACar'
+import SingleCarCard from '@/app/showcase/cars/inventory/SingleCarCard'
+import Success from '@/app/showcase/checkout/Success'
+import Wishlist from '@/app/showcase/wishlist/Wishlist'
 import React from 'react'
-import { Route, Routes } from 'react-router-dom'
+import { Navigate, Route, Routes } from 'react-router-dom'
+import NotFound from '@/app/extra/404'
 
 const AppRoutes: React.FC = () => {
   return (
     <Routes>
       <Route path="/" element={<Home />} />
+      <Route path="/404" element={<NotFound />} />
+      <Route path="*" element={<Navigate to="/404" replace />} />
       <Route path="/auth/login" element={<Login />} />
       <Route path="/auth/register" element={<Register />} />
       <Route path='/auth/verify-registration' element={<VerifyRegistation />} />
